refactor: 토큰 DTO 위치 변경
