+++ src/main/java/com/thinktank/api/dto/auth/TokenReqDto.java
@@ -1,4 +1,4 @@
-package com.thinktank.api.dto;
+package com.thinktank.api.dto.auth;
 
 public record TokenReqDto(
 	String expiredToken
